requirements.txt
requests==2.18.4
beautifulsoup4==4.6.0
tqdm==4.19.4
mecab-python3==0.7
wordcloud==1.5.0
matplotlib==3.0.0
pandas==0.23.4
gensim==3.6.0
